feature: rename `Result` type to `ResultFactory`
