+++ test/unit/Monads/Mothers/ResultMother.cs
@@ -5,11 +5,11 @@ internal static class ResultMother
 	internal static Result<Constellation, string> Default => default;
 
 	internal static Result<Constellation, string> Succeed()
-		=> Result.Succeed<Constellation, string>(ResultFixture.Success);
+		=> ResultFactory.Succeed<Constellation, string>(ResultFixture.Success);
 
 	internal static Result<Constellation, string> Succeed(Constellation success)
-		=> Result.Succeed<Constellation, string>(success);
+		=> ResultFactory.Succeed<Constellation, string>(success);
 
 	internal static Result<Constellation, string> Fail(string failure)
-		=> Result.Fail<Constellation, string>(failure);
+		=> ResultFactory.Fail<Constellation, string>(failure);
 }
